+++ src/woocommerce_hipayenterprise/includes/gateways/local/class-hipay-multibanco.php
@@ -95,8 +95,6 @@ class Hipay_Multibanco extends Hipay_Gateway_Local_Abstract
 
             $this->logs->logInfos("# Process Payment for $order_id");
 
-            $method = $this->confHelper->getLocalPayment($this->paymentProduct);
-
             $params = array(
                 "order_id" => $order_id,
                 "paymentProduct" => $this->paymentProduct,
@@ -160,21 +158,12 @@ class Hipay_Multibanco extends Hipay_Gateway_Local_Abstract
             'multibanco.php',
             'frontend',
             array(
-                    'entity' => $this->entity,
-                    'reference' => $this->reference,
-                    'amount' => $this->amount,
-                    'expirationDate' => $this->expirationDate,
-                    'logo' => $this->getMultibancoIconUrl(),
-                )
+                'entity' => $this->entity,
+                'reference' => $this->reference,
+                'amount' => $this->amount,
+                'expirationDate' => $this->expirationDate,
+                'sdkJsUrl' => $this->confHelper->getPaymentGlobal()["sdk_js_url"]
+            )
         );
-    }
-
-    /**
-     * Return Multibanco logo URL
-     * @return url
-     */
-    private function getMultibancoIconUrl()
-    {
-        return WC_HIPAYENTERPRISE_URL_ASSETS . 'local_payments_images/multibanco.png';
     }
 }