+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+- **Add** : Added Sisal reference without redirection
+- **Add** : Added Mooney reference without redirection
+- **Fix** : Fixed PHP 8 issue
+
 ## 2.3.1
 
 - **Fix**: Fixed 3DS v2 merchant risk information builder